Refactored moves and board utils
